+++ src/clj/skylobby/fx/mod_sync.clj
@@ -22,7 +22,8 @@
    #"Tech Annihilation"])
 
 (def no-rapid
-  [#"Total Atomization Prime"])
+  [#"Total Atomization Prime"
+   #"Evolution RTS Music Addon"])
 
 
 (defn mod-download-source [mod-name]
@@ -36,7 +37,7 @@
 
 (defn- mod-sync-pane-impl
   [{:fx/keys [context]
-    :keys [dependency engine-version mod-name spring-isolation-dir]}]
+    :keys [dependency engine-version index-only mod-name spring-isolation-dir]}]
   (let [copying (fx/sub-val context :copying)
         downloadables-by-url (fx/sub-val context :downloadables-by-url)
         file-cache (fx/sub-val context :file-cache)
@@ -49,7 +50,7 @@
         tasks-by-type (fx/sub-ctx context skylobby.fx/tasks-by-type-sub)
         indexed-mod (fx/sub-ctx context sub/indexed-mod spring-isolation-dir mod-name)
         mod-details (fx/sub-ctx context skylobby.fx/mod-details-sub indexed-mod)
-        no-mod-details (and (not dependency) (not (resource/details? mod-details)))
+        no-mod-details (not (resource/details? mod-details))
         refresh-mods-tasks (fx/sub-ctx context skylobby.fx/tasks-of-type-sub :spring-lobby/refresh-mods)
         mod-details-tasks (fx/sub-ctx context skylobby.fx/tasks-of-type-sub :spring-lobby/mod-details)
         update-download-sources (->> (fx/sub-ctx context skylobby.fx/tasks-of-type-sub :spring-lobby/update-downloadables)
@@ -77,9 +78,10 @@
      (concat
        (let [severity (if no-mod-details
                         (if indexed-mod
-                          -1 2)
-                        (if (or dependency (= mod-name (:mod-name indexed-mod)))
-                          0 1))]
+                          (if (or dependency index-only)
+                            0 -1)
+                          2)
+                        0)]
          [{:severity severity
            :text "info"
            :human-text mod-name
@@ -310,14 +312,17 @@
                                  (if in-progress
                                    (str "Resetting " (fs/filename (:file mod-details))
                                         " git to ref " battle-mod-git-ref)
-                                   (str "Reset " (fs/filename (:file mod-details))
-                                        " git to ref " battle-mod-git-ref)))
+                                   (if (seq mod-update-tasks)
+                                     "Refreshing mods..."
+                                     (str "Reset " (fs/filename (:file mod-details))
+                                          " git to ref " battle-mod-git-ref))))
                    :action
                    {:event/type :spring-lobby/add-task
                     :task
                     {:spring-lobby/task-type :spring-lobby/git-mod
                      :file mod-file
-                     :battle-mod-git-ref battle-mod-git-ref}}}))]
+                     :battle-mod-git-ref battle-mod-git-ref
+                     :spring-root spring-isolation-dir}}}))]
              (when (and (not (zero? severity))
                         (not= battle-mod-git-ref "$VERSION"))
                [(merge